# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Primavera V10 API - LLM Ready",
    "version": "4.0.0",
    "description": "API REST otimizada para consumo por LLMs. Dados do ERP Cegid Primavera V10 da Cimenteira do Louro. Todos os cálculos de faturação usam 34 tipos de documento com notas de crédito como valores negativos.",
    "contact": {
      "name": "Cimenteira do Louro",
      "url": "http://primavera-api.cimenteiradolouro.local"
    }
  },
  "servers": [
    {
      "url": "http://primavera-api.cimenteiradolouro.local",
      "description": "Servidor Primavera API"
    }
  ],
  "security": [
    { "basicAuth": [] },
    { "apiKeyHeader": [] },
    { "apiKeyQuery": [] }
  ],
...
  "paths": {
    "/api/clientes": {
      "get": {
        "tags": ["Clientes"],
        "summary": "Listar clientes",
        "description": "Retorna lista paginada de clientes com pesquisa e filtros por zona e vendedor.",
        "operationId": "listarClientes",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Pesquisa por nome, código ou contribuinte do cliente",
            "schema": { "type": "string" }
          },
          {
            "name": "zona",
            "in": "query",
            "description": "Filtrar por zona geográfica",
            "schema": { "type": "string" }
          },
          {
            "name": "vendedor",
            "in": "query",
            "description": "Filtrar por código do vendedor",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número da página (começa em 1)",
            "schema": { "type": "integer", "default": 1, "minimum": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número de resultados por página",
            "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 500 }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ClienteResumo" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/clientes/segmentacao": {
      "get": {
        "tags": ["Clientes"],
        "summary": "Segmentação de clientes",
        "description": "Análise de segmentação de clientes por zona e localidade, com métricas de faturação por período.",
        "operationId": "segmentacaoClientes",
        "parameters": [
          {
            "name": "zona",
            "in": "query",
            "description": "Filtrar por zona geográfica",
            "schema": { "type": "string" }
          },
          {
            "name": "localidade",
            "in": "query",
            "description": "Filtrar por localidade",
            "schema": { "type": "string" }
          },
          {
            "name": "periodo",
            "in": "query",
            "description": "Período de análise",
            "schema": { "type": "string", "enum": ["mes", "trimestre", "ano"], "default": "ano" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número máximo de resultados",
            "schema": { "type": "integer", "default": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados de segmentação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "zona": { "type": "string", "example": "Norte" },
                          "localidade": { "type": "string", "example": "Braga" },
                          "totalClientes": { "type": "integer", "example": 45 },
                          "faturacao": { "type": "number", "format": "double", "example": 125430.50 },
                          "percentagemTotal": { "type": "number", "format": "double", "example": 12.5 }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/clientes/top": {
      "get": {
        "tags": ["Clientes"],
        "summary": "Top clientes",
        "description": "Ranking dos melhores clientes por faturação no período especificado.",
        "operationId": "topClientes",
        "parameters": [
          {
            "name": "zona",
            "in": "query",
            "description": "Filtrar por zona geográfica",
            "schema": { "type": "string" }
          },
          {
            "name": "localidade",
            "in": "query",
            "description": "Filtrar por localidade",
            "schema": { "type": "string" }
          },
          {
            "name": "periodo",
            "in": "query",
            "description": "Período de análise",
            "schema": { "type": "string", "enum": ["mes", "trimestre", "ano"], "default": "ano" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número de clientes no ranking",
            "schema": { "type": "integer", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranking de clientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "posicao": { "type": "integer", "example": 1 },
                          "cliente": { "type": "string", "example": "C0001" },
                          "nome": { "type": "string", "example": "Construções Silva Lda" },
                          "zona": { "type": "string", "example": "Norte" },
                          "faturacao": { "type": "number", "format": "double", "example": 245000.00 },
                          "percentagemTotal": { "type": "number", "format": "double", "example": 8.2 }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/clientes/{id}": {
      "get": {
        "tags": ["Clientes"],
        "summary": "Detalhes do cliente",
        "description": "Retorna informação detalhada de um cliente específico incluindo dados cadastrais, contactos e condições comerciais.",
        "operationId": "detalheCliente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Código do cliente",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do cliente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/ClienteDetalhe" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/clientes/{id}/perfil-financeiro": {
      "get": {
        "tags": ["Clientes"],
        "summary": "Perfil financeiro do cliente",
        "description": "Retorna o perfil financeiro do cliente incluindo crédito disponível, dívida atual e valores em atraso.",
        "operationId": "perfilFinanceiroCliente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Código do cliente",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil financeiro",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cliente": { "type": "string", "example": "C0001" },
                        "nome": { "type": "string", "example": "Construções Silva Lda" },
                        "limiteCredito": { "type": "number", "format": "double", "example": 50000.00 },
                        "creditoDisponivel": { "type": "number", "format": "double", "example": 32500.00 },
                        "dividaTotal": { "type": "number", "format": "double", "example": 17500.00 },
                        "dividaVencida": { "type": "number", "format": "double", "example": 3200.00 },
                        "diasAtrasoMedio": { "type": "integer", "example": 15 },
                        "documentosPendentes": { "type": "integer", "example": 4 },
                        "ultimoPagamento": { "type": "string", "format": "date", "example": "2026-02-28" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Autenticação HTTP Basic"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Chave de API via header"
      },
      "apiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey",
        "description": "Chave de API via query parameter"
      }
    },
    "schemas": {
      "ClienteResumo": {
        "type": "object",
        "properties": {
          "cliente": { "type": "string", "example": "C0001" },
          "nome": { "type": "string", "example": "Construções Silva Lda" },
          "contribuinte": { "type": "string", "example": "501234567" },
          "zona": { "type": "string", "example": "Norte" },
          "localidade": { "type": "string", "example": "Braga" },
          "vendedor": { "type": "string", "example": "V001" },
          "telefone": { "type": "string", "example": "+351 253 123 456" },
          "email": { "type": "string", "example": "geral@construcoessilva.pt" },
          "ativo": { "type": "boolean", "example": true }
        }
      },
      "ClienteDetalhe": {
        "type": "object",
        "properties": {
          "cliente": { "type": "string", "example": "C0001" },
          "nome": { "type": "string", "example": "Construções Silva Lda" },
          "nomeComercial": { "type": "string", "example": "Silva Construções" },
          "contribuinte": { "type": "string", "example": "501234567" },
          "morada": { "type": "string", "example": "Rua do Comércio, 123" },
          "localidade": { "type": "string", "example": "Braga" },
          "codigoPostal": { "type": "string", "example": "4700-001" },
          "zona": { "type": "string", "example": "Norte" },
          "pais": { "type": "string", "example": "PT" },
          "telefone": { "type": "string", "example": "+351 253 123 456" },
          "telemovel": { "type": "string", "example": "+351 912 345 678" },
          "email": { "type": "string", "example": "geral@construcoessilva.pt" },
          "vendedor": { "type": "string", "example": "V001" },
          "nomeVendedor": { "type": "string", "example": "João Santos" },
          "condicoesPagamento": { "type": "string", "example": "30 dias" },
          "modoExpedicao": { "type": "string", "example": "Transporte próprio" },
          "limiteCredito": { "type": "number", "format": "double", "example": 50000.00 },
          "descontoFinanceiro": { "type": "number", "format": "double", "example": 2.0 },
          "ativo": { "type": "boolean", "example": true },
          "dataCriacao": { "type": "string", "format": "date", "example": "2018-03-15" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 50 },
          "total": { "type": "integer", "example": 1250 },
          "totalPages": { "type": "integer", "example": 25 },
          "hasNext": { "type": "boolean", "example": true },
          "hasPrev": { "type": "boolean", "example": false }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Não autorizado - credenciais inválidas ou em falta",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string", "example": "Autenticação necessária" },
                "code": { "type": "string", "example": "UNAUTHORIZED" }
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Erro interno do servidor",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string", "example": "Erro interno do servidor" },
                "code": { "type": "string", "example": "INTERNAL_ERROR" }
              }
            }
          }
        }
      }
    }
  }
}